# 6 семестр/Основы промышленной разработки программного обеспечения/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /api/user/:
    get:
      tags:
      - api
      - user
      summary: Get All Users
      operationId: get_all_users_api_user__get
      responses:
        '200':
          description: List all users
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
                title: Response Get All Users Api User  Get
    post:
      tags:
      - api
      - user
      summary: Create User
      operationId: create_user_api_user__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '201':
          description: Add new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/user/{user_id}:
    get:
      tags:
      - api
      - user
      summary: Get User
      operationId: get_user_api_user__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Retrieve a user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - api
      - user
      summary: Update User
      operationId: update_user_api_user__user_id__put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Update user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - api
      - user
      summary: Delete User
      operationId: delete_user_api_user__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '204':
          description: Delete a user
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateUser:
      properties:
        name:
          type: string
          title: Name
        surname:
          type: string
          title: Surname
        lastname:
          type: string
          title: Lastname
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - name
      - surname
      - lastname
      - email
      - password
      title: CreateUser
      example:
        email: iiivanov@edu.ru
        lastname: "\u0418\u0432\u0430\u043D\u043E\u0432\u0438\u0447"
        name: "\u0418\u0432\u0430\u043D"
        password: sdfsdfwgesdgcx
        surname: "\u0418\u0432\u0430\u043D\u043E\u0432"
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Role:
      type: string
      enum:
      - admin
      - regular
      title: Role
...
    UpdateUser:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        surname:
          anyOf:
          - type: string
          - type: 'null'
          title: Surname
        lastname:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastname
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        position:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      required:
      - name
      - surname
      - lastname
      - email
      - position
      - password
      title: UpdateUser
      example:
        email: iiivanov@edu.ru
        lastname: "\u0418\u0432\u0430\u043D\u043E\u0432\u0438\u0447"
        name: "\u0418\u0432\u0430\u043D"
        password: sdfsdfwgesdgcx
        position: regular
        surname: "\u0418\u0432\u0430\u043D\u043E\u0432"
    User:
      properties:
        _id:
          type: string
          title: ' Id'
        name:
          type: string
          title: Name
        surname:
          type: string
          title: Surname
        lastname:
          type: string
          title: Lastname
        email:
          type: string
          format: email
          title: Email
        position:
          $ref: '#/components/schemas/Role'
        password:
          type: string
          title: Password
      type: object
      required:
      - name
      - surname
      - lastname
      - email
      - position
      - password
      title: User
      example:
        email: iiivanov@edu.ru
        id: 42ea02b5-0e50-430f-bc68-aa6017122fde
        lastname: "\u0418\u0432\u0430\u043D\u043E\u0432\u0438\u0447"
        name: "\u0418\u0432\u0430\u043D"
        password: sdfsdfwgesdgcx
        position: regular
        surname: "\u0418\u0432\u0430\u043D\u043E\u0432"
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...